feat: ui-dashboard-stream-timeline-split-recovery-v2-006 - Dashboard stream/timeline shell regression proof
Split app-shell session-stream harness modules under Biome line limits and add app-shell refresh regression (loading shell, stream reopen, timeline reset).
